requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
chromadb==0.5.23
openai==1.58.1
anthropic==0.40.0
google-api-python-client==2.157.0
google-auth==2.37.0
python-dotenv==1.0.1
tiktoken==0.8.0
pydantic==2.10.4
